update attn_deocde: bf16, bench workload, cudnn

+++ ezops/kernels/attn_decode/attn_decode_cu.cu
@@ -1,14 +1,15 @@
 #include <tvm/ffi/tvm_ffi.h>
 #include <math.h>
+#include <cuda_bf16.h>
 
 // TODO: implement the CUDA kernel for attn_decode
 // attn_decode computes: out = softmax(Q @ K^T / sqrt(head_dim)) @ V
-// where Q is [batch, num_heads, 1, head_dim]
-//       K is [batch, num_heads, seq_len, head_dim]
-//       V is [batch, num_heads, seq_len, head_dim]
-//       out is [batch, num_heads, 1, head_dim]
+// where Q is [batch, num_heads, 1, head_dim]  (bfloat16)
+//       K is [batch, num_heads, seq_len, head_dim]  (bfloat16)
+//       V is [batch, num_heads, seq_len, head_dim]  (bfloat16)
+//       out is [batch, num_heads, 1, head_dim]  (bfloat16)
 __global__ void attn_decode_kernel(
-    const float* Q, const float* K, const float* V, float* Out,
+    const __nv_bfloat16* Q, const __nv_bfloat16* K, const __nv_bfloat16* V, __nv_bfloat16* Out,
     int batch, int num_heads, int seq_len, int head_dim,
     int stride_qb, int stride_qh, int stride_qm, int stride_qd,
     int stride_kb, int stride_kh, int stride_km, int stride_kd,
